Push up an actual release starting with version 1.3.0 since it's been 3 months since the last "release"
Not to many changes between the last "latest" and this one, I just think it was about time to get an actual new release out.

I will now be including versioning in the console log as the beginning to help differentiate versioning a tiny bit to help with debugging some stuff.

Also I have added a fix to now differentiate between SDH and regular Subtitles, mainly thanks to @dekaidekai for both the code and for testing it out for me.