Take 2 on trying to reduce CPU usage during the connected test.
Last attempt was to try and change the call to more a more lightweight call doing just a libtype search instead of fetching all sections, which seemed to be working okay at first, but eventually it started to run into errors (the main one being `Failed to get access token for {USER}`)

New solution seems to be working a lot better and I have tested for ~5 hours without any errors coming up.

It now caches the connection status (currently every 5 minutes but I might lower it a bit) and will use the cached result until the next check interval.

I will probably need to fiddle with things a bit more eventually especially for the `except (BadRequest, RequestsConnectionError, IndexError)` and  `except Exception` error stuff but I mainly am wanting to get this out there so I can get feedback on if it's going to work for others or not who have much more users than myself on their plex.