feat: try to improve audio stream title matching and filtering
#15

Hopefully helps with not choosing `visualImpaired` audio streams unless explicitly set (like commentary), right now it only checks for the term in `extendedDisplayTitle`, `displayTitle` and `title` as there doesn't seem to be a way to check for a `visualImpaired` attribute any.

I honestly don't even know if this is going to work or not as I've been in a fog this whole week due to being sick, and I do not have any video examples myself to test against this.