fix(track_changes): Handle missing season or episode numbers in range calculation
Not 100% sure on if the we need the `else` or not, as I wasn't able to get it to spit out `Unable to determine range due to missing season or episode number for` bit any but I'm going to leave it as a `if else` for now until I know for sure that it's not needed to be this way.

Also version 1.3.5 release!